requirements.txt
alembic==0.9.6
amqp==2.2.2
asn1crypto==0.23.0
billiard==3.5.0.2
celery==4.1.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
click==6.7
cryptography==2.1.4
Flask==0.12.2
Flask-Cors==3.0.3
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
kombu==4.1.0
Mako==1.0.7
MarkupSafe==1.0
psycopg2==2.7.3.2
pycparser==2.18
pyOpenSSL==17.5.0
PySocks==1.6.7
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.3
PyYAML==3.12
redis==2.10.6
requests==2.18.4
simplejson==3.11.1
six==1.11.0
SQLAlchemy==1.1.13
urllib3==1.22
vine==1.1.4
Werkzeug==0.12.2